mlops-cv-pipeline/requirements.txt
# Training
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24
pillow>=10.0
pyyaml>=6.0

# Experiment tracking / registry
mlflow>=2.9.0

# Serving
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9

# Monitoring
prometheus-client>=0.19

# Dev/test
pytest>=8.0
httpx>=0.26
ruff>=0.3

# Optional (drift reports)
# evidently>=0.4.20
